cplay: combine speed incr, decr, reset methods
All three methods contained almost identical code and are now combined in a new
method incr_decr_reset_speed() that takes the signum of the change as a parameter.

This patch also changes the key binding of the speed reset from '=' to the
previously unused '\' and therefore reverts the binding of '=' to its old
function 'volume increment'. This reduces confusion and increases the intuition
of the key bindings.