cplay: Reordered player definition
To honour the original concept of Ulf Betlehem of a very lightweight console
music player the definition of players was changed. MPlayer was moved further
to the back so that players consuming less system resources have priority if
any of them is installed (see README).
Individual preferences can be set in /etc/cplayrc or ~/.cplayrc (use "make
cplayrc" to create an example file).